pyproject.toml
[project]
name = "salescoach"
version = "0.1.0"
description = "AI Sales Coach Platform"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.14.0",
    "pydantic>=2.9.2",
    "pydantic-settings>=2.6.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "httpx>=0.27.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
]
audio = [
    "whisper>=1.0.0",
    "openai-whisper>=20240930",
    "pyannote.audio>=3.3.1",
    "sounddevice>=0.5.2",
    "soundfile>=0.12.1",
    "pydub>=0.25.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
